refactor: Rename IStatusService to IActivityService and move its dependencies to the automation namespace
