+++ src/PiWeb.Import.Sdk/Modules/ImportAutomation/IActivityService.cs
@@ -8,31 +8,14 @@
 
 #endregion
 
-using Zeiss.PiWeb.Import.Sdk.Activity;
-
 namespace Zeiss.PiWeb.Import.Sdk.Modules.ImportAutomation;
 
 /// <summary>
 /// Responsible for managing and displaying the current status (activity and recent events) of an
 /// <see cref="IImportRunner"/>. A status service is usually connected to the user interface of the hosting application.   
 /// </summary>
-public interface IStatusService
+public interface IActivityService
 {
-    /// <summary>
-    /// Posts a new entry to the event list. A list of the latest events of an import plan is shown on the status page
-    /// of the Auto Importer UI.
-    /// </summary>
-    /// <param name="displayText">
-    /// The display text of the event. A localization handler will be used to localize this text with the given
-    /// arguments. Implement <see cref="IPlugin.CreateLocalizationHandler"/> to specify your own localization and
-    /// formatting.
-    /// </param>
-    /// <param name="formatArgs">
-    /// Format arguments used when formatting the display text of the activity.
-    /// </param>
-    /// <param name="severity">The severity of the event.</param>
-    public void PostImportEvent( EventSeverity severity, string displayText, params object[] formatArgs );
-
     /// <summary>
     /// Sets the current activity. The current activity of an import plan is shown at various places of the
     /// Auto Importer UI either in the form of the detailed text or the short text. Other properties of the current
@@ -54,4 +37,19 @@ public interface IStatusService
     /// Resets the current activity to no activity.
     /// </summary>
     public void ClearActivity();
+    
+    /// <summary>
+    /// Posts a new entry to the list of recent events of an import plan. The event list of an import plan is shown
+    /// on the status page of the Auto Importer UI.
+    /// </summary>
+    /// <param name="displayText">
+    /// The display text of the event. A localization handler will be used to localize this text with the given
+    /// arguments. Implement <see cref="IPlugin.CreateLocalizationHandler"/> to specify your own localization and
+    /// formatting.
+    /// </param>
+    /// <param name="formatArgs">
+    /// Format arguments used when formatting the display text of the activity.
+    /// </param>
+    /// <param name="severity">The severity of the event.</param>
+    public void PostActivityEvent( EventSeverity severity, string displayText, params object[] formatArgs );
 }
